+++ 99_outputs.tf
@@ -1,4 +1,4 @@
-#### IoT Thing ####
+### IoT ###
 output "iot_thing_default_client_id" {
   value = aws_iot_thing.poc.default_client_id
 }
@@ -11,13 +11,10 @@ output "iot_thing_arn" {
     value = aws_iot_thing.poc.arn
 }
 
-#### IoT Thing Type ####
-
 output "iot_thing_type_arn" {
     value = aws_iot_thing_type.poc.arn
 }
 
-#### Certificate ####
 output "iot_certificate_id" {
     value = aws_iot_certificate.poc.id
 }
@@ -33,4 +30,13 @@ output "iot_certificate_public_key" {
 }
 output "iot_certificate_private_key" {
     value = aws_iot_certificate.poc.private_key
+}
+output "elastisearch_domain_arn" {
+    value = aws_elasticsearch_domain.poc.arn
+}
+output "elastisearch_domain_id" {
+    value = aws_elasticsearch_domain.poc.domain_id
+}
+output "elastisearch_domain_endpoint" {
+    value = aws_elasticsearch_domain.poc.endpoint
 }