+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes from aws-iot-tfg will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 15-02-2020
+
+- IoT Thing, Type and certificate
+- Non-functional Elasticsearch service
+- Outputs and variables
+
 ## [0.1.0] - 15-02-2020
 
 - Main Terraform file.